+++ internal/provider/provider.go
@@ -1588,9 +1588,10 @@ func Provider() *schema.Provider {
 			"aws_mskconnect_custom_plugin":        kafkaconnect.ResourceCustomPlugin(),
 			"aws_mskconnect_worker_configuration": kafkaconnect.ResourceWorkerConfiguration(),
 
-			"aws_kendra_faq":       kendra.ResourceFaq(),
-			"aws_kendra_index":     kendra.ResourceIndex(),
-			"aws_kendra_thesaurus": kendra.ResourceThesaurus(),
+			"aws_kendra_faq":                          kendra.ResourceFaq(),
+			"aws_kendra_index":                        kendra.ResourceIndex(),
+			"aws_kendra_query_suggestions_block_list": kendra.ResourceQuerySuggestionsBlockList(),
+			"aws_kendra_thesaurus":                    kendra.ResourceThesaurus(),
 
 			"aws_keyspaces_keyspace": keyspaces.ResourceKeyspace(),
 			"aws_keyspaces_table":    keyspaces.ResourceTable(),