+++ internal/service/kendra/find.go
@@ -38,6 +38,33 @@ func FindFaqByID(ctx context.Context, conn *kendra.Client, id, indexId string) (
 	return out, nil
 }
 
+func FindQuerySuggestionsBlockListByID(ctx context.Context, conn *kendra.Client, id, indexId string) (*kendra.DescribeQuerySuggestionsBlockListOutput, error) {
+	in := &kendra.DescribeQuerySuggestionsBlockListInput{
+		Id:      aws.String(id),
+		IndexId: aws.String(indexId),
+	}
+
+	out, err := conn.DescribeQuerySuggestionsBlockList(ctx, in)
+	if err != nil {
+		var resourceNotFoundException *types.ResourceNotFoundException
+
+		if errors.As(err, &resourceNotFoundException) {
+			return nil, &resource.NotFoundError{
+				LastError:   err,
+				LastRequest: in,
+			}
+		}
+
+		return nil, err
+	}
+
+	if out == nil {
+		return nil, tfresource.NewEmptyResultError(in)
+	}
+
+	return out, nil
+}
+
 func FindThesaurusByID(ctx context.Context, conn *kendra.Client, id, indexId string) (*kendra.DescribeThesaurusOutput, error) {
 	in := &kendra.DescribeThesaurusInput{
 		Id:      aws.String(id),